multierror: move to internal package
